+++ app/Main.hs
@@ -1,7 +1,8 @@
 module Main (main) where
 
-import Lib
+import Core
+import Parser
 
 main :: IO ()
-main = someFunc
+main = putStr "Resultado: " >> ((putStrLn . show) $ parseFormula "p an q" 1 "stdin")